// src/types/flow.ts
export interface TraitCapabilities {
    add?: string[];
    drop?: string[];
}

export interface TraitSELinuxOptions {
    level?: string;
    role?: string;
    type?: string;
    user?: string;
}

export interface TraitSeccompProfile {
    type: 'Unconfined' | 'RuntimeDefault' | 'Localhost';
    localhostProfile?: string;
}

export interface TraitAppArmorProfile {
    type: 'Unconfined' | 'RuntimeDefault' | 'Localhost';
    localhostProfile?: string;
}

export interface TraitSecurityPolicy {
    name: string;
    capabilities?: TraitCapabilities;
    privileged?: boolean;
    seLinuxOptions?: TraitSELinuxOptions;
    runAsUser?: number;
    runAsGroup?: number;
    runAsNonRoot?: boolean;
    readOnlyRootFilesystem?: boolean;
    allowPrivilegeEscalation?: boolean;
    procMount?: 'Default' | 'Unmasked';
    seccompProfile?: TraitSeccompProfile;
    appArmorProfile?: TraitAppArmorProfile;
}